The recipe search is fonctionnal, add search for diet regimes and tolerable. The composant is independant and make a call to the api by itself
