+++ src/components/RecipeCard.js
@@ -2,18 +2,17 @@ import React from 'react'
 import axios from 'axios'
 import RecipeSearch from './RecipeSearch'
 import DisplayRecipe from './DisplayRecipe'
-import GetRecipeApiDatas from './getRecipeApiDatas'
 
 class RecipeCard extends React.Component {
 
   state = {
     recipe: '',
-    ingredients: [],
-    ingredient1: "",
-    ingredient2: "",
-    ingredient3: "",
-    errorIngredient: "",
-    userCalories: 0
+    ingredients: []
+    // ingredient1: "",
+    // ingredient2: "",
+    // ingredient3: "",
+    // errorIngredient: "",
+    // userCalories: 0
   }
 
   getRecipe(ingredient1, ingredient2, ingredient3, userCalories) {
@@ -37,19 +36,16 @@ class RecipeCard extends React.Component {
         // Search for the recipe whom match with the max of calories
         const objectUri = res1.filter(x => x.recipe.calories === thisCalories)[0]
         // Define the state with the research recipe and the ingredients which go with
-        console.log(res.data.count, "numberRecipeFound")
-        console.log(randomNum, "randomNum")
         this.setState({
           recipe: objectUri.recipe,
           ingredients: objectUri.recipe.ingredientLines,
-          ingredient1: '',
-          ingredient2: '',
-          ingredient3: '',
-          errorIngredient: ""
+          // ingredient1: '',
+          // ingredient2: '',
+          // ingredient3: '',
+          // errorIngredient: ""
         })
-
       })
-      .catch(error => this.setState({ errorIngredient: "Erreur dans la saie des ingrédients" }))
+    // .catch(error => this.setState({ errorIngredient: "Erreur dans la saie des ingrédients" }))
 
   }
 
@@ -62,32 +58,32 @@ class RecipeCard extends React.Component {
     return time > 60 ? `${hours} ${unity} and ${minutes} minutes` : `${minutes} minutes`
   }
 
-  submitForm = (e) => {
-    e.preventDefault()
-    this.getRecipe(this.state.ingredient1, this.state.ingredient2, this.state.ingredient3, this.state.userCalories)
-    console.log(this.state.ingredient1, "ingredient1")
-    console.log(this.state.ingredient2, "ingredient2")
-    console.log(this.state.ingredient3, "ingredient3")
-  }
+  // submitForm = (e) => {
+  //   e.preventDefault()
+  //   this.getRecipe(this.state.ingredient1, this.state.ingredient2, this.state.ingredient3, this.state.userCalories)
+  //   console.log(this.state.ingredient1, "ingredient1")
+  //   console.log(this.state.ingredient2, "ingredient2")
+  //   console.log(this.state.ingredient3, "ingredient3")
+  // }
   getOtherRecipe = () => {
     this.getRecipe()
   }
 
-  handleChange = (e) => {
-    let userIngredient1 = e.target.id === "firstIngredient" ? e.target.value : this.state.ingredient1
-    let userIngredient2 = e.target.id === "secondIngredient" ? e.target.value : this.state.ingredient2
-    let userIngredient3 = e.target.id === "thirdIngredient" ? e.target.value : this.state.ingredient3
-    let actualUserCalories = e.target.id === "actualUserCalories" ? e.target.value : this.state.userCalories
+  // handleChange = (e) => {
+  //   let userIngredient1 = e.target.id === "firstIngredient" ? e.target.value : this.state.ingredient1
+  //   let userIngredient2 = e.target.id === "secondIngredient" ? e.target.value : this.state.ingredient2
+  //   let userIngredient3 = e.target.id === "thirdIngredient" ? e.target.value : this.state.ingredient3
+  //   let actualUserCalories = e.target.id === "actualUserCalories" ? e.target.value : this.state.userCalories
 
 
-    this.setState({
-      ingredient1: userIngredient1,
-      ingredient2: userIngredient2,
-      ingredient3: userIngredient3,
-      userCalories: actualUserCalories,
-    })
-    // console.log(this.state.ingredient3)
-  }
+  //   this.setState({
+  //     ingredient1: userIngredient1,
+  //     ingredient2: userIngredient2,
+  //     ingredient3: userIngredient3,
+  //     userCalories: actualUserCalories,
+  //   })
+  //   // console.log(this.state.ingredient3)
+  // }
 
 
 
@@ -101,14 +97,8 @@ class RecipeCard extends React.Component {
 
     return (
       <div className="RecipeCard" >
-        <RecipeSearch
-          handleChange={this.handleChange}
-          submitForm={this.submitForm}
-          errorIngredient={this.state.errorIngredient}
-          actualUserCalories={this.state.userCalories}
-        />
         <DisplayRecipe getOtherRecipe={this.getOtherRecipe} ingredientsList={this.state.ingredients} recipe={this.state.recipe} preparationTime={this.getPreparationTime(totalTime)} calories={calories} />
-        <GetRecipeApiDatas ingredients="cheese" calories={100} />
+        <RecipeSearch />
       </div>
     );
   }
